kb/cicd/cicd_jenkins_project_deployer/requirements.txt
dataiku-api-client >=9.0.0
pandas==0.23.4
radon
pytest
